test application.conf is now recognized removing deprecated class
