+++ event-log-verification-service/src/main/scala/com/ubirch/verification/service/util/Exceptions.scala
@@ -1,7 +1,7 @@
 package com.ubirch.verification.service.util
 
 import com.ubirch.util.Exceptions.ExecutionException
-import com.ubirch.verification.service.process.LookupPipeDataNew
+import com.ubirch.verification.service.eventlog.LookupPipeDataNew
 
 /**
   * Namespace that contains the exceptions of the SDK.
